Add start of Learning Tasks, add FileType
