+++ shared/src/commonMain/kotlin/org/orca/kotlass/data/activity/ActivityInstance.kt
@@ -1,4 +1,4 @@
-package org.orca.kotlass.data.common
+package org.orca.kotlass.data.activity
 
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.SerialName
@@ -8,67 +8,9 @@ import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.booleanOrNull
 import kotlinx.serialization.json.jsonObject
 import kotlinx.serialization.json.jsonPrimitive
-
-/**
- * Compass representation of a given activity, which has a list
- * of instances. Activities are generally a [Class], but others exist.
- */
-@Serializable(with = Activity.ActivitySerializer::class)
-sealed interface Activity {
-
-    /**
-     * Name of this activity.
-     */
-    @SerialName("ActivityDisplayName")
-    val name: String
-
-    /**
-     * Unique ID of this activity.
-     */
-    @SerialName("ActivityId")
-    val id: Int
-
-    /**
-     * List of [ActivityInstance]s belonging to this Activity.
-     */
-    @SerialName("Instances")
-    val instances: List<ActivityInstance>
-
-    /**
-     * A given Class instance which has lessons.
-     */
-    @Serializable
-    data class Class(
-        @SerialName("ActivityDisplayName")
-        override val name: String,
-
-        @SerialName("ActivityId")
-        override val id: Int,
-
-        @SerialName("Instances")
-        override val instances: List<ActivityInstance.ClassInstance>,
-
-        /**
-         * Readable name of this subject.
-         */
-        @SerialName("SubjectName")
-        val subjectName: String,
-    ) : Activity
-
-    /**
-     * Compass differentiates activities by a series of "`Is...`" boolean properties.
-     */
-    object ActivitySerializer : JsonContentPolymorphicSerializer<Activity>(Activity::class) {
-        override fun selectDeserializer(element: JsonElement): DeserializationStrategy<out Activity> =
-            element.jsonObject.let {
-                if (it["IsStandardClass"]?.jsonPrimitive?.booleanOrNull == true) {
-                    return Class.serializer()
-                }
-
-                throw IllegalArgumentException("Activity belongs to no activity type!")
-            }
-    }
-}
+import org.orca.kotlass.data.common.fileasset.FileAsset
+import org.orca.kotlass.data.common.LocationContainer
+import org.orca.kotlass.data.common.Manager
 
 @Serializable(with = ActivityInstance.ActivityInstanceSerializer::class)
 sealed interface ActivityInstance {
@@ -142,19 +84,3 @@ sealed interface ActivityInstance {
             }
     }
 }
-
-/**
- * Request for getting a given [Activity] by its [Activity.id].
- */
-@Serializable
-internal data class CompassGetActivityById(
-    val activityId: Int
-)
-
-/**
- * Request for getting a given [Activity] by an [instanceId] belonging to it.
- */
-@Serializable
-internal data class CompassGetActivityByInstanceId(
-    val instanceId: String
-)
